+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     * `block::SetAttribute` is a modifier which overrides a single attribute of a block, instead of all of them.
       It replaces `Modifier::Attributes`.
     * `math::Aab::expand_or_shrink()`
+    * `math::FaceMap::{as_array, as_array_mut}()`
     * `math::Vol::origin_empty()`, infallible construction of empty volume.
     * `physics::Contact::aab()`
     * `time::Clock::phase()`, returns the clock’s phase, which was previously only obtainable through `Tick`.